+++ src/main/java/com/sparklenote/user/oAuth2/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.sparklenote.user.oAuth;
+package com.sparklenote.user.oAuth2;
 
 import com.sparklenote.user.dto.response.UserResponseDTO;
 import lombok.RequiredArgsConstructor;
